Rename: likes -> like로 수정
