+++ src/main/java/com/teamddd/duckmap/dto/event/like/LikeRes.java
@@ -1,4 +1,4 @@
-package com.teamddd.duckmap.dto.event.likes;
+package com.teamddd.duckmap.dto.event.like;
 
 import lombok.Builder;
 import lombok.Getter;
